Initialize playerScore and computerScore. Create game function. Make playRound function keep score
Created variables playerScore and computerScore to keep score of game.
game() function uses a for loop to call playRound() 5 times. Each time playRound is called either playerScore or computerScore is incremented by 1 for the winner. game() also compares playerScore to computerScore after the loop and console.log's a winner of the game.